requirements.txt
fastapi==0.110.0
uvicorn
httpx==0.27.0 #Gemini gibi dış servislerle haberleşme
pyrebase4
python-dotenv
google-generativeai
streamlit==1.45.0
pydantic==2.6.4  # FastAPI 0.110 ile uyumlu, v1 değil!
pydantic[email]